apps/clothing-processor/pyproject.toml
[tool.poetry]
name = "clothing-processor"
version = "0.1.0"
description = "API to process and upload clothing items."
authors = ["leo.mosley <leo.mosley2@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
test-api = "clothing_processor.scripts.test_api:main"
push-env = "clothing_processor.scripts.push_env:main"
test-colour-match = "clothing_processor.scripts.test_colour_match:main"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
absl-py = "2.1.0"
alembic = "1.14.1"
amqp = "5.3.1"
annotated-types = "0.7.0"
anyio = "4.8.0"
astunparse = "1.6.3"
attrs = "25.1.0"
authlib = "1.4.1"
automaton = "3.2.0"
autopage = "0.5.2"
awscli = "1.37.8"
awslambdaric = "3.0.1"
bcrypt = "4.2.1"
boto3 = "1.36.8"
botocore = "1.36.8"
cachetools = "5.5.1"
certifi = "2024.12.14"
cffi = "1.17.1"
charset-normalizer = "3.4.1"
click = "8.1.8"
cliff = "4.8.0"
cmd2 = "2.5.11"
colorama = "0.4.6"
coloredlogs = "15.0.1"
cryptography = "44.0.0"
debtcollector = "3.0.0"
decorator = "5.1.1"
dnspython = "2.7.0"
docker = "7.1.0"
docutils = "0.16"
dogpile-cache = "1.3.4"
email-validator = "2.2.0"
eventlet = "0.39.0"
fastapi = "0.115.7"
fastapi-cli = "0.0.7"
fasteners = "0.19"
flatbuffers = "25.1.24"
futurist = "3.0.0"
gast = "0.6.0"
google-pasta = "0.2.0"
greenlet = "3.1.1"
grpcio = "1.70.0"
h11 = "0.14.0"
h5py = "3.12.1"
httpcore = "1.0.7"
httptools = "0.6.4"
httpx = "0.28.1"
humanfriendly = "10.0"
idna = "3.10"
imageio = "2.37.0"
importlib-metadata = "8.6.1"
iso8601 = "2.1.0"
jinja2 = "3.1.5"
jmespath = "1.0.1"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
keras = "3.8.0"
keystoneauth1 = "5.9.1"
keystonemiddleware = "10.8.0"
kombu = "5.4.2"
lazy-loader = "0.4"
libclang = "18.1.1"
llvmlite = "0.44.0"
logutils = "0.3.5"
mako = "1.3.8"
mangum = "0.19.0"
markdown = "3.7"
markdown-it-py = "3.0.0"
markupsafe = "3.0.2"
mdurl = "0.1.2"
ml-dtypes = "0.4.1"
mpmath = "1.3.0"
msgpack = "1.1.0"
namex = "0.0.8"
netaddr = "1.3.0"
networkx = "3.4.2"
numba = "0.61.0"
numpy = "2.0.2"
onnxruntime = "1.20.1"
opencv-python-headless = "4.11.0.86"
openstacksdk = "4.3.0"
opt-einsum = "3.4.0"
optree = "0.14.0"
os-client-config = "2.1.0"
os-service-types = "1.7.0"
osc-lib = "3.2.0"
oslo-cache = "3.9.0"
oslo-concurrency = "7.0.0"
oslo-config = "9.7.0"
oslo-context = "5.7.0"
oslo-db = "17.1.0"
oslo-i18n = "6.5.0"
oslo-log = "7.0.0"
oslo-messaging = "16.0.0"
oslo-metrics = "0.10.1"
oslo-middleware = "6.3.0"
oslo-policy = "4.5.0"
oslo-reports = "3.5.0"
oslo-serialization = "5.6.0"
oslo-service = "4.0.0"
oslo-upgradecheck = "2.4.0"
oslo-utils = "8.0.0"
oslo-versionedobjects = "3.5.0"
packaging = "24.2"
paste = "3.10.1"
pastedeploy = "3.1.0"
pbr = "6.1.0"
pecan = "1.5.1"
pillow = "11.1.0"
platformdirs = "4.3.6"
pooch = "1.8.2"
prettytable = "3.13.0"
prometheus-client = "0.21.1"
protobuf = "5.29.3"
psutil = "6.1.1"
pyasn1 = "0.6.1"
pycadf = "4.0.1"
pycparser = "2.22"
pydantic = "2.10.6"
pydantic-core = "2.27.2"
pydot = "3.0.4"
pygments = "2.19.1"
pyjwt = "2.10.1"
pymatting = "1.1.13"
pyopenssl = "25.0.0"
pyparsing = "3.2.1"
pyperclip = "1.9.0"
pyreadline3 = "3.5.4"
python-barbicanclient = "7.0.0"
python-cinderclient = "9.6.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
python-glanceclient = "4.7.0"
python-heatclient = "4.1.0"
python-keystoneclient = "5.5.0"
python-multipart = "0.0.20"
python-neutronclient = "11.4.0"
python-novaclient = "18.8.0"
python-octaviaclient = "3.9.0"
python-openstackclient = "7.2.1"
python-swiftclient = "4.6.0"
pytz = "2024.2"
pywin32 = "308"
pyyaml = "6.0.2"
referencing = "0.36.2"
rembg = "2.0.61"
repoze-lru = "0.7"
requests = "2.32.3"
requestsexceptions = "1.4.0"
rfc3986 = "2.0.0"
rich = "13.9.4"
rich-toolkit = "0.13.2"
routes = "2.5.1"
rpds-py = "0.22.3"
rsa = "4.7.2"
s3transfer = "0.11.2"
scikit-image = "0.25.1"
scipy = "1.15.1"
shellingham = "1.5.4"
simplegeneric = "0.8.1"
simplejson = "3.19.3"
six = "1.17.0"
snapshot-restore-py = "1.0.0"
sniffio = "1.3.1"
sqlalchemy = "2.0.37"
starlette = "0.45.3"
statsd = "4.0.1"
stevedore = "5.4.0"
sympy = "1.13.3"
taskflow = "5.10.0"
tenacity = "9.0.0"
tensorboard = "2.18.0"
tensorboard-data-server = "0.7.2"
tensorflow = "2.18.0"
tensorflow-io-gcs-filesystem = "0.31.0"
tensorflow-intel = "2.18.0"
termcolor = "2.5.0"
testresources = "2.0.1"
testscenarios = "0.5.0"
testtools = "2.7.2"
tifffile = "2025.1.10"
tqdm = "4.67.1"
typer = "0.15.1"
typing-extensions = "4.12.2"
tzdata = "2025.1"
urllib3 = "2.3.0"
uvicorn = "0.34.0"
vine = "5.1.0"
warlock = "2.0.1"
watchfiles = "1.0.4"
wcwidth = "0.2.13"
webob = "1.8.9"
websockets = "14.2"
werkzeug = "3.1.3"
wrapt = "1.17.2"
wsme = "0.12.1"
yappi = "1.6.10"
zipp = "3.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
